Add RailsTag Component
This components takes in html generated by all kinds of rails helper
and creates a react component with the outer most element, then use
dangerouslySetHtml for the contents of that element. The advantages
of this is that we no longer need to declare a empty div or span to
house these elemnents while adding onclick handlers direclty to the
tags.