Allow customization of visit and remote
This moves the enhancement of the visit to application.js and passes to the
NavComponent. Now whenever a page component gets rendered, the custom
dispatch gets passed instead of the original visit.

The value of this becomes more clear when a user adds before and after effects
when visiting the next page. For example, showing and clearing a loader before
and after a visit.