Add support for redux persist
We add generators for redux persist. This solves an issue with loading the page
from disk cache and having all modifications disappear, and the application
loading again from its initial state set in the HEAD tag.

This is solved by introducing Redux-Persist. I've been mulling over this
decision because its another layer to worry about. Saving the state on the
client side is dangerous because the JS assets depends on data that was shaped
at that moment in time. Migrations on the client side is complexity that I do
not want, ever.

I came around to it when I realize that we can avoid migrations by setting the
local storage key name to the JS assets. This is a far more simplier solution.
Now every time assets change, we get a new key in local storage and ignore the
old keys.