Enhance PropsTemplate to directly call Oj
This commit removes the need for PropsTemplate to create a command queue before
`send`ing the to Oj. This means we call Oj directly without creating an
intermediate data stucture!

We also add an improvement to caching to remove to serialize and deserialize
from the cache store. This is achieved by storing a new line character in the
cache between the meta data and the JSON payload. When we fetch, we deserialize
only the first line, but the JSON payload is passed as a string onto Oj's
`push_json`.

In adding these changes, I've also borrowed the performance testing suite from
TurboStreamer to benchmark this library against various other implementations.
The results were insightful:

1. PropsTemplate performs the fastest among all implementations including
Turbostreamer which uses StreamWriter instead of StringWriter. In terms of
memory usage, it is comparable with rabl. Need more speed? Props::Base is the
base class that PropsTemplate uses, and is twice aw fast as PropsTemplate.

2. In terms of caching, PropsTemplate outperforms popular library like active
model serializers, jbuilder, and rabl by nearly 4x. Not surprising given the
ability to cache without deserializing. However, it falls a little behind
Turbostreamer.

I've also added a few improvements to reduce memory allocations.
1. A key cache has been added to Props::Base
2. Only a single instance of Oj::StringWriter is used per process. This
optimization is needed because performance degrades significantly on multiple
calls to Stringwriter's `new`. If you use a forking server like Puma, be sure to
call Props.reset_encoder! on_worker_boot.

These are breaking changes.