Fix missing x-xhr-referrer origin
The upgrade for url-parse introduced a breaking change, if the second param to parse
is undefined it will use window.location. We're passing an empty has to keep the
behavior somewhat the same with the nodejs version. But now this breaks the way
x-xhr-referrer works since it no longer included the origin portion of the url. So
we make the baseURL explicit and default it to the `window.location.url`.

I wonder if we really need x-xhr-referrer, it was ported from Turbolinks 3, more
investigation needed.