Add fragment support
A fragment is marked node indicating that it is a cross cutting concern across
all pages. For example, a header that appears across multiple pages. This is
similar to the idea of a [Turbo-Frame](https://turbo.hotwired.dev/handbook/frames).

When a new page loads with a fragment Breezy will look at the page store and update
nodes marked with the same fragment name across all pages.

This is also true for fragments modified by a custom reducer. If you use a custom
reducer to update the page state, fragments of the same name across all pages will
also update. This behavior is controlled by a new middleware. This is included by
default when using generating a breezy app:

```
import { ApplicationBase, fragmentMiddleware } from '@jho406/breezy'

...
composeEnhancers(applyMiddleware(thunk, fragmentMiddleware))
```

You can turn this behavior off by removing the middleware.