Remove redux persist
Redux persist isn't needed. It was included originally to handle the case of a
direct visit or back button visit to a page that had the redux store mutated
from the original injected state.

But I realized I could just lean on the [cache header] support from rails to
handle freshness and force the browser to give us a new page, with the newly
injected JSON too.

[cache header]: https://thoughtbot.com/blog/take-control-of-your-http-caching-in-rails