Remove Rails tag.
Superglue now has its own JSON form helper [form_props], that stands in as the
default uncontrolled input builder. When used with Humid, there's no longer a
need to render rails form as HTML, parse it as html with html-react-renderer
and rerender as react. We can just use the form props as props to the React
input elements.