Refactor busters (#28)
* Remove anticache

This functionality isn't needed as we can handle cache control headers on the
server side.

* Replace __=0 with format=json

The original intent of __=0 was to bust the browser cache that happens
when visiting a page in HTML format where subsequent visits to the same
page with a different content negotiation would return the wrong format.

I did not know of Rail's own `format=json` that you can add the param to help
prevent that issue from ever happening. This replaces that.