Add warning for remote reqs
This adds a warning to remote when it tries to replace a page that has a different
componentIdentifier that a received page. This can happen when you redirect to a
completely new page using the current page's url via a form submit. And since remote
doesn't navigate away from the current page, the current page component may receive
props that are shaped differently.