Remove virtual_path_of_template (#32)
* Remove virtual_path_of_template

This helper is being removed in favor of "#{controller_path}/#{action_name}". There's
a small difference between both approaches. With the change now, we're able to render
the same template with different componentIdentifier for example

```ruby
def show
  render :edit
end
```

This would result in the same template rendered, but would have a different
componentIdentifier. Useful for cases like modals.

* Remove unused require