Use form_props active_template_virtual_path
This reverts an earlier removal of of the method `virtual_path_of_template`.
Unfortunately, a combination of `controller_path` and `action_name` caused
some unintended consequences of some actions rendering the wrong template.

Using `controller_path` and `action_name` would've been a nice way to DRY
things up, but we're going to revert for now until we find a better method.