Add support for remote forms using GET
When encountering a GET or HEAD request, superglue explicitly
removes the body. That's because the native fetch would throw
an error if a body would ever appear on a GET. On a normal
form however, a form with a GET would result in the form data
being embedded in the URL.

This replicates that behavior by turning the form data into URL
search params.