Fix issues 406 errors
We set content-type null on UJS as a flag to remove it later down
the stack. I'm not sure why I did it this way, but it missed a case
when the method type on the form was a GET. This resulted in
content-type of "null" being sent to the server.

This isn't the best fix and I have to find a better way of achieving
the same thing.