Simplify the setting of pageKey
This commit makes Superglue a bit smarter. Previously, content-location was
used to specify the pageKey for 200s from non-get requests, but this was too
cumbersome as the most common use case was setting the URL to be the referrer
of the non-get request upon 200.

This uses that as the default logic but still accepts content-location if
we want to explicitly override from the server, which is unlikely. Additionally
I simplified the logic. We always want to use the response's URL unless in
the case mentioned above. Existing tests pass.