Simplify fragments
Fragments has been experimental for some time. The main idea of it was to
provide a way to update things that appeared across pages or if it appeared
multiple time in the same page.

But something smelled about it, especially the added complexity and the numerous
what ifs.

So instead this PR simplifies fragments down to the following: Whenever
Superglue encounters a fragment during a visit, or if the fragment was
subsequently updated because of a deferment, we fire `UPDATE_FRAGMENT`.

That's it. We don't mutate state in response to that action. Instead we allow
the users to make use of the action in their own slices. Typically to update
the slice state.