Make placeholder smarter
Makes the determination of the placeholder value default to the currentPageKey.
This reflects the majority of use cases for placeholder, where we want to
selectively update the current page a user is viewing while changing the URL.
Modals is a good example.

This also removes the data-placeholder data-attribute option. There is feature
work in the pipeline which would give greater flexibility to allow a user to add
back the functionality.