Expose url search string as an object
And add it to the navigation context. This takes inspiration
from Emberjs, where the URL is a reflection of application state.
For most apps, we end up accessing the query params somehow on the
frontend anyway. This adds a convenience to have it available in the
navigation context.

It can be used as a lighter weight `useState` where it is important
for the user to revisit the same page and get the same application
frontend.