Fix hash navigations
If we land on a superglue enabled page click on a hash, then data-sg-visit to a new
superglue page, an error would be thrown. Superglue ignored hash scenarios because it
using hashes shouldn't affect the app, but apparently it does with this edge case.

To resolve this, any hash url now will have a location.state.superglue as long as the
pagekey exists in our pages store.