Use react lifecycle to manage scroll
This fixes a issue where scroll restoration was not working as it was called before
react renders the restored page. In this change we move the restoration to
`useLayoutEffect`.

In reviewing the fix, we noticed that `setActivePage` was not being dispatched
in `navigateTo`, and relied on `onHistoryChange` to make the correct changes.
We refactor to move the explicit nav logic from `onHistoryChange` and back to
`navigateTo`.