Introduce next iteration of useContent
This version of useContent takes fragments into account and instead of
returning the object from redux, it returns a proxy of the current page.

The proxy is lazy, and at every encounter of a proxy ref `{__id: 'abc-123'}`
it will look up the fragments in the story, proxy it and return that. This
makes accessing the entire page seamless while still getting the benefits
of having fragments with identities across the application.

The dependency tracking is build around redux's useSelector. At every fragment
access through the proxy, we store that id in the useContent hook, this tracks
the dependencies.  We're using useRef for this so it does trigger any updates.
When Redux state changes, all useSelector hook are rerun and we use the compare
function to run through the dependencies by using === on the prev state and
next state.

An `unproxy` method is provided to return the value behind the proxy. This is
useful for referential equality when needed and preventing needless dependency tracking

for example

```
const {user} = useContent()

//assuming profile is a fragment
// and i pass it to a child component

<Profile profile={user.profile}>

```

When accessing profile, the fragment id is added to the dependency list of
the current component and if the only thing i do with a profile is pass it to
a child this means that if the profile changes, the parent rerenders and the
child rerenders, but if i unproxy it before hand.

```
const {user} = useContent()
<Profile profileRef={unproxy(user).profile}>
```

the dependency doesn't get tracked on the parent component.The child component
can do

```
useFragment(profileRef)
```

And when profile gets updated, only the profile component gets rerendered and
not the parent.