Add useSetFragment to replace useFragment
All of the useFragment retrieval usecases is covered by `useContent`, so we don't need
the getter of useFragment. useSetFragment represents the setter, what's unique is that
it returns a function that you use to wrap another function and that will return a
draft that you can modify. for example:

```
const set = useSetFragment()

set('cart', (cart) => {
  cart.lineItems.push(...)

 // if cart.footer was another fragment
  set(cart.footer, (footer) => (footer.total = 19))
})
```

each set will dispatch a saveFragment action in redux. This serves as an easier
alternative to mutate state than setting up slices, etc.