Add Fragment type.
This helps the `unproxy` method to return a type without specifying it. For example:

```
const page = useContent<{
  header: Fragment<{
    name: string;
  }>;
}>();
```

then when you use `unproxy(page)`. The type would be

```
{
  header: {
    __id: string;
  };
}
```

We also fix rename the old Fragment type to FragmentRef, and change `type` to `id`,
which is what superglue_rails will return in v2