Fix useContent edge cases
This fixes a few things with useContent

1. Remove circular weakmaps. We had a local weakmap with a redux value pointing to a proxy
and a global weakmap with a proxy pointing to a redux value. While I could probably use
WeakRef in the global, I opted for a more simplier solution using Symbols that allows me
to return the proxies target.

2. fragment was being curried which cause issues with fetching the actual fragment from
the state. Instead we pass a ref and access fragments via `current`.