Fix proxies issues.
1. `beforeSave` now is passed a proxy. This is useful because using the Prev page has
been normalized, making it difficult to manipulate the next set of results. By passing
a proxy we get the convenience of a denormalized page without the fragment refs.

2. The proxy cache is now created in useMemo. Previously it cached older objects that
became stale if a new fragment was saved with the same id.

3. Add `useStore` to always use the latest fragments when creating the proxy. Without
it the proxies passed from useSelector are incomplete. We rename that to `trackedFragments`
because its still used as a way to minimize rerenders when any existing fragment usage changes.
This change is about new incoming fragments.