Fix accidental frozen obj
When using `useSetFragment`, the obj returned was frozen. This cause a javascript
invariant violation which breaks the proxy leading to this error discovered on

https://github.com/thoughtbot/superglue/pull/173

```
Uncaught TypeError: 'get' on proxy: property 'toggleForm' is a read-only and non-configurable data property on the proxy target but the proxy did not return its actual value (expected '#<Object>' but got '#<Object>')
    at Item (ItemsList.jsx:11:5)
    at Object.react_stack_bottom_frame (react-dom-client.development.js:23863:20)
... etc ..
```

The fix is to use a local instance of Immer and setting the auto freeze to false.